feature AB#72151 fixed large assetFile elastic index, allow to delete custom distribution. Flush tmp files on kernel.terminate event.

+++ src/Resources/config/elasticsearch/asset.php
@@ -67,7 +67,7 @@ return static function (ContainerConfigurator $configurator): void {
                     'type' => 'keyword',
                 ],
                 'size' => [
-                    'type' => 'integer',
+                    'type' => 'long',
                 ],
                 'ratioWidth' => [
                     'type' => 'integer',
